+++ packages/core/src/index.ts
@@ -1,2 +1,1 @@
 export { Web3InboxClient } from "./client";
-export * from "./state/snapshots";
