+++ packages/react/src/hooks/useManageSubscription.ts
@@ -1,5 +1,5 @@
 import type { NotifyClientTypes } from "@walletconnect/notify-client";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { ErrorOf, HooksReturn, LoadingOf, SuccessOf } from "../types/hooks";
 import { useWeb3InboxClient } from "./useWeb3InboxClient";
